+++ playground/CHANGELOG.md
@@ -1,5 +1,47 @@
 # playground
 
+## 0.0.68
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@3.1.10
+
+## 0.0.67
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@3.1.9
+
+## 0.0.67
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@3.1.8
+
+## 0.0.66
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@3.1.7
+
+## 0.0.65
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@3.1.6
+
+## 0.0.64
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@3.1.5
+
 ## 0.0.63
 
 ### Patch Changes